Edited Query Package
Created an expression that exports the results of running that query to a CSV file in Query 3, 4, 6, 7, 9, 10. As well as ran those queries to create CSV files of their results. Those files were saved to the maindb/resources folder. Also reran Query 2 to fix an accidental overwrite of that query.